requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
click==8.1.8
coloredlogs==15.0.1
distro==1.9.0
flatbuffers==25.2.10
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.8.2
livekit==0.20.1
livekit-agents==0.12.15
livekit-api==0.8.2
livekit-plugins-openai==0.11.0
livekit-plugins-silero==0.7.4
livekit-protocol==0.9.0
mpmath==1.3.0
multidict==6.1.0
numpy==2.2.3
onnxruntime==1.20.1
openai==1.64.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
sympy==1.13.3
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
watchfiles==0.24.0
yarl==1.18.3
